complete page 수정

+++ complete.js
@@ -1,13 +1,18 @@
 const hello = document.getElementById("hello");
 const logoutButton = document.getElementById("logout");
-const name = Cookies.get("username");
-hello.innerText = `안녕하세요. ${name}님!`;
+const name = localStorage.getItem("ID");
 
-logoutButton.onclick = function () {
-  localStorage.clear();
-  sessionStorage.clear();
-  Cookies.remove("ID");
-  Cookies.remove("username");
-  Cookies.remove("loginflag");
+if (localStorage.getItem("loginFlag") == "logincheck") {
+  hello.innerText = `안녕하세요. ${name}님!`;
+
+  logoutButton.onclick = function () {
+    localStorage.clear();
+    sessionStorage.clear();
+    Cookies.remove("ID");
+    Cookies.remove("username");
+    Cookies.remove("loginflag");
+    window.location.replace("index.html");
+  };
+} else {
   window.location.replace("index.html");
-};
+}
